+++ apps/docs/[doc].paths.ts
@@ -1,8 +1,8 @@
-import type { BadgeTree } from '@readme-widget-hub/badge'
 import type { Feature } from '@readme-widget-hub/render'
+import type { WidgetTree } from '@readme-widget-hub/widget'
 import path from 'node:path'
 import { Manager } from '@readme-widget-hub/manager'
-import { renderBadgeDoc, renderIndexDoc } from '@readme-widget-hub/render'
+import { renderIndexDoc, renderWidgetDoc } from '@readme-widget-hub/render'
 import { path2Url } from '@readme-widget-hub/utils'
 import { loadEnv } from './.vitepress/env'
 
@@ -17,28 +17,28 @@ const rootDir = path.resolve(__dirname, '../../')
 const env = loadEnv(rootDir)
 const manager = new Manager({
   defaultLocaleCode: env.VITE_DEFAULT_LOCALE_CODE,
-  absBadgesDir: path.join(rootDir, env.VITE_BADGES_DIR),
+  absWidgetsDir: path.join(rootDir, env.VITE_WIDGETS_DIR),
   absMetaFilePath: path.join(rootDir, env.VITE_META_FILE_PATH),
 })
 const locales = manager.getLocales()
 
-function badgeFilePath2Url(localeUrlPrefix: string, filePath: string): string {
+function widgetFilePath2Url(localeUrlPrefix: string, filePath: string): string {
   return path2Url(path.join(localeUrlPrefix, path.relative(rootDir, filePath))).slice(0, -5)
 }
 
-function badge2Features(localeUrlPrefix: string, flatBadges: BadgeTree[]): Feature[] {
+function widget2Features(localeUrlPrefix: string, flatWidgets: WidgetTree[]): Feature[] {
   const features: Feature[] = []
   let count = 0
   const showCount = Number.parseInt(env.VITE_DOC_SHOW_FEATURES_COUNT)
-  for (const badge of flatBadges) {
+  for (const widget of flatWidgets) {
     if (count >= showCount) {
       break
     }
-    if (badge.type === 'badge') {
+    if (widget.type === 'widget') {
       features.push({
-        title: badge.title,
-        details: badge.description,
-        link: badgeFilePath2Url(localeUrlPrefix, badge.path),
+        title: widget.title,
+        details: widget.description,
+        link: widgetFilePath2Url(localeUrlPrefix, widget.path),
       })
       count++
     }
@@ -53,7 +53,7 @@ export default {
     for (const locale of locales) {
       const localeUrlPrefix = manager.isDefaultLocale(locale) ? '' : locale.code
       const meta = manager.getMeta(locale)
-      const flatBadges = manager.getFlatBadges(locale)
+      const flatWidgets = manager.getFlatWidgets(locale)
 
       // index
       paths.push({
@@ -62,21 +62,21 @@ export default {
           name: meta.name,
           title: meta.title,
           logo: '/logo.svg',
-          features: badge2Features(localeUrlPrefix, flatBadges),
+          features: widget2Features(localeUrlPrefix, flatWidgets),
           doc: meta.doc,
           localeCode: localeUrlPrefix,
         }),
       })
 
-      for (const flatBadge of flatBadges) {
-        if (flatBadge.type !== 'badge') {
+      for (const flatWidget of flatWidgets) {
+        if (flatWidget.type !== 'widget') {
           continue
         }
 
         paths.push({
-          params: { doc: badgeFilePath2Url(localeUrlPrefix, flatBadge.path) },
-          content: renderBadgeDoc({
-            badge: flatBadge,
+          params: { doc: widgetFilePath2Url(localeUrlPrefix, flatWidget.path) },
+          content: renderWidgetDoc({
+            widget: flatWidget,
             readme: meta.readme,
             showParams: env.VITE_SHOW_PARAMS === 'true',
           }),