+++ apps/readme/src/index.ts
@@ -1,14 +1,14 @@
 import path from 'node:path'
 import { Manager } from '@readme-widget-hub/manager'
-import { renderBadge, renderReadme } from '@readme-widget-hub/render'
+import { renderReadme, renderWidget } from '@readme-widget-hub/render'
 import fs from 'fs-extra'
-import { badgeFilePath2Url, badgeTree2Tocs, object2Navs } from './transform'
+import { object2Navs, widgetFilePath2Url, widgetTree2Tocs } from './transform'
 
 const rootDir = path.resolve(__dirname, '../../../')
 const env = import.meta.env
 const manager = new Manager({
   defaultLocaleCode: env.VITE_DEFAULT_LOCALE_CODE,
-  absBadgesDir: path.join(rootDir, env.VITE_BADGES_DIR),
+  absWidgetsDir: path.join(rootDir, env.VITE_WIDGETS_DIR),
   absMetaFilePath: path.join(rootDir, env.VITE_META_FILE_PATH),
 })
 const meta = manager.getMeta()
@@ -17,7 +17,7 @@ for (const locale of meta.locales) {
   const localeMeta = manager.getMeta(locale)
   const localeFileName = manager.locale2FileName(locale)
   const readme = localeMeta.readme
-  const flatBadges = manager.getFlatBadges(locale)
+  const flatWidgets = manager.getFlatWidgets(locale)
 
   const outputPath = path.join(rootDir, manager.locale2FileName(locale))
   console.log(`Generating ${outputPath}`)
@@ -28,35 +28,35 @@ for (const locale of meta.locales) {
     title: localeMeta.title,
     description: localeMeta.description,
     showTags: true,
-    badgeCount: manager.badgeCount,
+    widgetCount: manager.widgetCount,
     navs: object2Navs(false, readme, readme, env.VITE_ONLINE_PAGE_URL, locale, manager),
-    tocs: badgeTree2Tocs(rootDir, flatBadges, env, localeFileName),
+    tocs: widgetTree2Tocs(rootDir, flatWidgets, env, localeFileName),
     readme,
   })
 
   fs.mkdirSync(path.dirname(outputPath), { recursive: true })
   fs.writeFileSync(outputPath, markdown)
 
-  for (const flatBadge of flatBadges) {
-    if (flatBadge.type === 'collection') {
+  for (const flatWidget of flatWidgets) {
+    if (flatWidget.type === 'collection') {
       continue
     }
 
-    const badgeUrl = badgeFilePath2Url(rootDir, flatBadge.path, env, localeFileName)
-    const outputPath = path.join(rootDir, badgeUrl)
-    const badgeData = manager.getBadge(flatBadge.path, locale)
+    const widgetUrl = widgetFilePath2Url(rootDir, flatWidget.path, env, localeFileName)
+    const outputPath = path.join(rootDir, widgetUrl)
+    const widgetData = manager.getWidget(flatWidget.path, locale)
     console.log(`Generating ${outputPath}`)
 
-    const markdown = renderBadge({
+    const markdown = renderWidget({
       mode: 'markdown',
-      hasLocale: manager.hasLocale(flatBadge, locale),
+      hasLocale: manager.hasLocale(flatWidget, locale),
       title: localeMeta.title,
       description: localeMeta.description,
       showTags: false,
-      badgeCount: manager.badgeCount,
-      navs: object2Navs(true, badgeData, readme, env.VITE_ONLINE_PAGE_URL, locale, manager),
+      widgetCount: manager.widgetCount,
+      navs: object2Navs(true, widgetData, readme, env.VITE_ONLINE_PAGE_URL, locale, manager),
       readme: localeMeta.readme,
-      badge: badgeData,
+      widget: widgetData,
       showParams: env.VITE_SHOW_PARAMS === 'true',
     })