+++ apps/readme/src/transform.ts
@@ -1,39 +1,39 @@
-import type { BadgeTree } from '@readme-widget-hub/badge'
 import type { Manager } from '@readme-widget-hub/manager'
 import type { Locale, Readme } from '@readme-widget-hub/meta'
 import type { Nav, Toc } from '@readme-widget-hub/render'
+import type { WidgetTree } from '@readme-widget-hub/widget'
 import path from 'node:path'
 import { path2Url, replaceParentPath } from '@readme-widget-hub/utils'
 
-// badges/...json => /dist/badges/...md
-export function badgeFilePath2Url(rootDir: string, filePath: string, env: ImportMetaEnv, localeFileName: string): string {
+// widgets/...json => /dist/widgets/...md
+export function widgetFilePath2Url(rootDir: string, filePath: string, env: ImportMetaEnv, localeFileName: string): string {
   const { dir, name } = path.parse(replaceParentPath(
     filePath,
-    path.join(rootDir, env.VITE_BADGES_DIR),
-    env.VITE_BADGE_README_OUTPUT_DIR,
+    path.join(rootDir, env.VITE_WIDGETS_DIR),
+    env.VITE_WIDGET_README_OUTPUT_DIR,
   ))
   return path2Url(`/${path.join(dir, name, localeFileName)}`).slice(1)
 }
 
-// dist/badges/...md => badges/...json
-export function url2BadgeFilePath(rootDir: string, url: string, env: ImportMetaEnv): string {
+// dist/widgets/...md => widgets/...json
+export function url2WidgetFilePath(rootDir: string, url: string, env: ImportMetaEnv): string {
   const { dir } = path.parse(url)
-  const { dir: badgeDir, name } = path.parse(replaceParentPath(
+  const { dir: widgetDir, name } = path.parse(replaceParentPath(
     dir,
-    env.VITE_BADGE_README_OUTPUT_DIR,
-    env.VITE_BADGES_DIR,
+    env.VITE_WIDGET_README_OUTPUT_DIR,
+    env.VITE_WIDGETS_DIR,
   ))
-  return path.join(rootDir, badgeDir, `${name}.json`)
+  return path.join(rootDir, widgetDir, `${name}.json`)
 }
 
-export function badgeTree2Tocs(rootDir: string, flatBadgeTree: BadgeTree[], env: ImportMetaEnv, localeFileName: string): Toc[] {
-  return flatBadgeTree.map((node) => {
-    if (node.type === 'badge') {
+export function widgetTree2Tocs(rootDir: string, flatWidgetTree: WidgetTree[], env: ImportMetaEnv, localeFileName: string): Toc[] {
+  return flatWidgetTree.map((node) => {
+    if (node.type === 'widget') {
       return {
-        type: 'badge',
+        type: 'widget',
         name: node.title,
         level: node.level,
-        href: badgeFilePath2Url(rootDir, node.path, env, localeFileName),
+        href: widgetFilePath2Url(rootDir, node.path, env, localeFileName),
       }
     }
     return {